@peakji 2016-12-19
Heres the benchmark script, please ignore the nasty semicolons <del>(the pride and joy of a C programmer)</del> 😜

```javascript
const leveldown = require("/Users/PeakJi/Gits/leveldown");
const DBPATH = __dirname + "/tempdb";

var targets = [];
var db = leveldown(DBPATH);
db.open(function(err) {
	var batch = db.batch();
	for (var i = 0; i < 0xFFFFF; i++) {
		var buf = Buffer.allocUnsafe(4);
		buf.writeUInt32BE(i, 0, true);
		if (i > 0 && i % 1000 == 0)
			targets.push(buf);
		batch.put(buf, buf);
	}
	batch.write(function(err) {
		console.time("SKIP");
		skip(function(err) {
			console.timeEnd("SKIP");
			db.close(function(err) {
				leveldown.destroy(DBPATH, function(err) {

				});
			});
		});
	});
});

function scan(end) {
	var ite = db.iterator();
	ite.next(function iterate(err, key, val) {
		if (key)
			ite.next(iterate);
		else {
			ite.end(function(err) {
				end();
			});
		}
	});
}

function skip(end) {
	var ite = db.iterator();
	var run = function(i) {
		if (i == targets.length) {
			ite.end(function(err) {
				end();
			});
			return;
		}
		ite.seek(targets[i]);
		ite.next(function(err, key, val) {
			run(i + 1);
		});
	};
	run(0);
}
```