@ralphtheninja 2016-12-19
> Now we can do really really fast skip scan, sorted set intersection and K-way merging, with multiple seeks from #323.

What's K-way merging?