@peakji 2016-12-19
@ralphtheninja Its actually called [K-Way Merge Algorithms](https://en.wikipedia.org/wiki/K-Way_Merge_Algorithms) or K-way merge sort, useful in external sorting and inverted index.