@ralphtheninja 2018-06-20
@peakji I'm just going through the code again. If I understand correctly, the first `.next()` you do after a `.seek()` will ignore `highWaterMark` and not buffer the results in order to optimize `.seek()`, `.next()`, `.seek()`, `.next()` etc, but once you do `.seek()`, `.next()`, `.next()`, `.next()` etc then you will buffer. Correct?